:dizzy: move /group to routes/group_create.js
