+++ lib/utils.js
@@ -5,7 +5,7 @@ const crypto = require('crypto');
 module.exports = {
 
 	check_config: () => {
-		if (! fs.existsSync(`${__dirname}/configg.js`)) {
+		if (! fs.existsSync(`${__dirname}/../config.js`)) {
 			console.log('Please set up config.js');
 			process.exit();
 		}
