+++ routes/home.js
@@ -1,5 +1,7 @@
 const express = require('express');
 const router = express.Router();
+
+const config = require('../config');
 const db = require('../lib/db');
 const utils = require('../lib/utils');
